pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "repoinspo"
version = "0.1.0"
description = "Analyze repositories, find similar projects, and surface portable ideas."
readme = "PLAN.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Raj Contractor" },
]
dependencies = [
  "faiss-cpu>=1.8.0",
  "gitingest>=0.1.4",
  "httpx>=0.27.2",
  "litellm>=1.55.0",
  "mcp[cli]>=1.2.0",
  "pydantic>=2.10.0",
  "pydantic-settings>=2.6.1",
  "rich>=13.9.4",
  "typer>=0.15.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
  "pytest-asyncio>=0.24.0",
  "ruff>=0.8.4",
]

[project.urls]
Repository = "https://github.com/contractorr/repoinspo"

[project.scripts]
repoinspo = "repoinspo.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/repoinspo"]

[tool.hatch.build.targets.sdist]
include = [
  "/src",
  "/tests",
  "/PLAN.md",
  "/LICENSE",
  "/.env.example",
]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
  "integration: tests that call real external services",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
  "E",
  "F",
  "I",
  "UP",
  "B",
  "ASYNC",
]
